+++ lib/trackless/trackless_failure.dart
@@ -1,3 +1,6 @@
+import 'package:flutter/material.dart';
+import 'package:trackless/app_localizations.dart';
+
 class TracklessFailure {
   final int code;
   final int detailCode;
@@ -7,5 +10,15 @@ class TracklessFailure {
   @override
   String toString() {
     return ' (Code: ${this.code} - ${this.detailCode})';
+  }
+
+  /// Displays the failure using a snackbar
+  displayFailure(BuildContext context) {
+    Scaffold.of(context).showSnackBar(new SnackBar(
+      content: Text(
+          AppLocalizations.of(context).translate('error_${this.code}') +
+              this.toString()),
+      backgroundColor: Colors.red,
+    ));
   }
 }