+++ changelog.md
@@ -3,7 +3,8 @@
 ### New in this version
 
 * Added a global app state
-* Able to show basic infomation
+* Able to show basic account infomation
+* Localstorage is now working for account page
 
 ### Changed in this version